+++ core-services/beneficiary-idgen/src/main/java/org/egov/id/producer/IdGenProducer.java
@@ -1,0 +1,21 @@
+package org.egov.id.producer;
+
+import lombok.extern.slf4j.Slf4j;
+import org.egov.tracer.kafka.CustomKafkaTemplate;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+
+@Service
+@Slf4j
+public class IdGenProducer {
+
+    @Autowired
+    private CustomKafkaTemplate<String, Object> kafkaTemplate;
+
+    public void push(String topic, Object value) {
+        kafkaTemplate.send(topic, value);
+        log.debug("Message successfully sent to topic: {}", topic);
+    }
+}
+