+++ health-services/household/src/main/java/org/egov/household/repository/HouseholdMemberRepository.java
@@ -1,15 +1,17 @@
 package org.egov.household.repository;
 
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.exception.ExceptionUtils;
 import org.egov.common.data.query.builder.GenericQueryBuilder;
 import org.egov.common.data.query.builder.QueryFieldChecker;
 import org.egov.common.data.query.builder.SelectQueryBuilder;
-import org.egov.common.data.query.exception.QueryBuilderException;
 import org.egov.common.data.repository.GenericRepository;
 import org.egov.common.models.core.SearchResponse;
 import org.egov.common.models.household.HouseholdMember;
+import org.egov.common.models.household.Relationship;
 import org.egov.common.models.household.HouseholdMemberSearch;
 import org.egov.common.producer.Producer;
+import org.egov.household.repository.rowmapper.RelationshipRowMapper;
 import org.egov.household.repository.rowmapper.HouseholdMemberRowMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.redis.core.RedisTemplate;
@@ -18,6 +20,7 @@ import org.springframework.stereotype.Repository;
 import org.springframework.util.ReflectionUtils;
 
 import java.lang.reflect.Method;
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -25,19 +28,25 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 
 import static org.egov.common.utils.CommonUtils.constructTotalCountCTEAndReturnResult;
+import static org.egov.common.utils.CommonUtils.getIdList;
 import static org.egov.common.utils.CommonUtils.getIdMethod;
 
 @Repository
 @Slf4j
 public class HouseholdMemberRepository extends GenericRepository<HouseholdMember> {
 
+    private final RelationshipRowMapper relationshipRowMapper;
+
     @Autowired
     protected HouseholdMemberRepository(Producer producer,
                                         NamedParameterJdbcTemplate namedParameterJdbcTemplate,
                                         RedisTemplate<String, Object> redisTemplate,
                                         SelectQueryBuilder selectQueryBuilder,
-                                        HouseholdMemberRowMapper householdMemberRowMapper) {
+                                        HouseholdMemberRowMapper householdMemberRowMapper,
+                                        RelationshipRowMapper relationshipRowMapper
+    ) {
         super(producer, namedParameterJdbcTemplate, redisTemplate, selectQueryBuilder, householdMemberRowMapper, Optional.of("household_member"));
+        this.relationshipRowMapper = relationshipRowMapper;
     }
 
 
@@ -85,23 +94,32 @@ public class HouseholdMemberRepository extends GenericRepository<HouseholdMember
         paramsMap.put("offset", offset);
 
         List<HouseholdMember> householdMembers = this.namedParameterJdbcTemplate.query(queryBuilder.toString(), paramsMap, this.rowMapper);
+        fetchAndSetHouseholdMemberRelationship(householdMembers, includeDeleted);
 
         return SearchResponse.<HouseholdMember>builder().totalCount(totalCount).response(householdMembers).build();
     }
 
     public SearchResponse<HouseholdMember> findById(List<String> ids, String columnName, Boolean includeDeleted) {
-        List<HouseholdMember> objFound = findInCache(ids).stream()
-                .filter(entity -> entity.getIsDeleted().equals(includeDeleted))
-                .collect(Collectors.toList());
-        if (!objFound.isEmpty()) {
-            Method idMethod = getIdMethod(objFound, columnName);
-            ids.removeAll(objFound.stream()
-                    .map(obj -> (String) ReflectionUtils.invokeMethod(idMethod, obj))
-                    .collect(Collectors.toList()));
-            if (ids.isEmpty()) {
-                log.info("all objects were found in the cache, returning objects");
-                return SearchResponse.<HouseholdMember>builder().response(objFound).build();
+        List<HouseholdMember> objFound = new ArrayList<>();;
+        try {
+            objFound = findInCache(ids);
+            if (!includeDeleted) {
+                objFound = objFound.stream()
+                        .filter(entity -> !entity.getIsDeleted())
+                        .collect(Collectors.toList());
             }
+            if (!objFound.isEmpty()) {
+                Method idMethod = getIdMethod(objFound, columnName);
+                ids.removeAll(objFound.stream()
+                        .map(obj -> (String) ReflectionUtils.invokeMethod(idMethod, obj))
+                        .toList());
+                if (ids.isEmpty()) {
+                    log.info("all objects were found in the cache, returning objects");
+                    return SearchResponse.<HouseholdMember>builder().response(objFound).build();
+                }
+            }
+        } catch (Exception e) {
+            log.info("Error occurred while reading from cache {}", ExceptionUtils.getStackTrace(e));
         }
 
         String query = String.format("SELECT * FROM household_member where %s IN (:ids) AND isDeleted = false", columnName);
@@ -112,6 +130,7 @@ public class HouseholdMemberRepository extends GenericRepository<HouseholdMember
         paramMap.put("ids", ids);
 
         objFound.addAll(this.namedParameterJdbcTemplate.query(query, paramMap, this.rowMapper));
+        fetchAndSetHouseholdMemberRelationship(objFound, includeDeleted);
         putInCache(objFound);
         log.info("returning objects from the database");
         return SearchResponse.<HouseholdMember>builder().response(objFound).build();
@@ -123,7 +142,8 @@ public class HouseholdMemberRepository extends GenericRepository<HouseholdMember
         Map<String, Object> paramMap = new HashMap();
         paramMap.put("individualId", individualId);
         List<HouseholdMember> householdMembers = this.namedParameterJdbcTemplate.query(query, paramMap, this.rowMapper);
-        return SearchResponse.<HouseholdMember>builder().totalCount(Long.valueOf(householdMembers.size())).response(householdMembers).build();
+        fetchAndSetHouseholdMemberRelationship(householdMembers, false);
+        return SearchResponse.<HouseholdMember>builder().totalCount((long) householdMembers.size()).response(householdMembers).build();
     }
 
     public SearchResponse<HouseholdMember> findIndividualByHousehold(String householdId, String columnName) {
@@ -138,5 +158,39 @@ public class HouseholdMemberRepository extends GenericRepository<HouseholdMember
         List<HouseholdMember> householdMembers = this.namedParameterJdbcTemplate.query(query, paramMap, this.rowMapper);
 
         return SearchResponse.<HouseholdMember>builder().totalCount(totalCount).response(householdMembers).build();
+    }
+
+    private void fetchAndSetHouseholdMemberRelationship(List<HouseholdMember> householdMembers, Boolean includeDeleted) {
+        if (householdMembers.isEmpty()) {
+            return;
+        }
+        List<String> householdMemberIds = getIdList(householdMembers);
+        Map<String, List<Relationship>> idToObjMap = fetchRelationships(householdMemberIds, includeDeleted);
+        householdMembers.forEach(member -> member.setMemberRelationships(idToObjMap.get(member.getId())));
+    }
+
+    private Map<String, List<Relationship>> fetchRelationships(List<String> selfIds, Boolean includeDeleted) {
+        Map<String, Object> resourceParamsMap = new HashMap<>();
+        StringBuilder resourceQuery = new StringBuilder("SELECT * FROM household_member_relationship WHERE selfId IN (:selfIds) ");
+        resourceParamsMap.put("selfIds", selfIds);
+        if (Boolean.FALSE.equals(includeDeleted)) {
+            resourceQuery.append("AND isDeleted=:isDeleted");
+            resourceParamsMap.put("isDeleted", false);
+        }
+        List<Relationship> relationships = this.namedParameterJdbcTemplate.query(resourceQuery.toString(), resourceParamsMap,
+                this.relationshipRowMapper);
+        Map<String, List<Relationship>> idToObjMap = new HashMap<>();
+
+        relationships.forEach(relationship -> {
+            String memberId = relationship.getSelfId();
+            if (idToObjMap.containsKey(memberId)) {
+                idToObjMap.get(memberId).add(relationship);
+            } else {
+                List<Relationship> memberRelationships = new ArrayList<>();
+                memberRelationships.add(relationship);
+                idToObjMap.put(memberId, memberRelationships);
+            }
+        });
+        return idToObjMap;
     }
 }