+++ health-services/individual/src/main/java/org/egov/individual/service/EnrichmentService.java
@@ -1,10 +1,6 @@
 package org.egov.individual.service;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
-import java.util.UUID;
+import java.util.*;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
@@ -205,7 +201,10 @@ public class EnrichmentService {
         log.info("enriching individual id in identifiers");
         List<Identifier> identifiers = individual.getIdentifiers();
         if (identifiers != null) {
-            identifiers.forEach(identifier -> identifier.setIndividualId(individual.getId()));
+            identifiers.forEach(identifier -> {
+                identifier.setIndividualId(individual.getId());
+                identifier.setIndividualClientReferenceId(individual.getClientReferenceId());
+            });
             individual.setIdentifiers(identifiers);
         }
         return individual;
@@ -233,6 +232,7 @@ public class EnrichmentService {
                     .clientReferenceId(UUID.randomUUID().toString())
                     .identifierType(SYSTEM_GENERATED)
                     .identifierId(individual.getId())
+                    .individualClientReferenceId(individual.getClientReferenceId())
                     .build());
             individual.setIdentifiers(identifiers);
         }