+++ health-services/referralmanagement/src/main/java/org/egov/referralmanagement/config/ReferralManagementConfiguration.java
@@ -115,4 +115,16 @@ public class ReferralManagementConfiguration {
     @Value("${egov.mdms.search.endpoint}")
     private String mdmsSearchUrl;
 
+    @Value("${egov.service.request.downsync.enabled:false}")
+    private Boolean serviceRequestDownsyncEnabled;
+
+    @Value("${egov.service.request.search.batch.size:100}")
+    private Integer serviceRequestSearchBatchSize;
+
+    @Value("${egov.service.request.host}")
+    private String serviceRequestHost;
+
+    @Value("${egov.service.request.service.search.url}")
+    private String serviceRequestServiceSearchUrl;
+
 }