+++ health-services/individual/CHANGELOG.md
@@ -1,5 +1,15 @@
 All notable changes to this module will be documented in this file.
 
+## 1.1.8 - 2025-06-11
+
+- Added `IndividualClientReferenceId` for Individual Identifiers
+- Updated integration with User service for individual operations
+
+## 1.1.7 - 2025-04-29 
+
+- Added validators for beneficiary ID validation in individual creation and update flows.
+- Added the necessary application properties for integration with the new Beneficiary ID generation service
+
 ## 1.1.6 - 2024-08-29 
 
  - Added `ExistentEntityValidator` fixes